Replace instances of Bitcoin/bitcoins/BTC
Use a function instead that returns the correct currency name.

Still some instances of Bitcoin being hardcoded, like with the Casascius
coins (there are no Casascius NMCs, so I don't know what to do) and in
the settings in relation to the torrent download (I don't think there is
a NMC torrent).